+++ packages/dara-core/js/components/index.tsx
@@ -9,5 +9,6 @@ export { default as CustomFallback } from './fallback/custom';
 export { default as For } from './for/for';
 export { default as Link } from './link';
 export { default as Outlet } from './outlet';
+export { default as Navigate } from './navigate';
 export { default as PoweredByCausalens } from './powered-by-causalens';
 export { default as ThemeProvider } from './theme-provider';